move to external, submodule geojsonvt
